playbook: Call to misc-perftools for compute file
Need to cause a file transfer for configuring telegraf
for compute hosts so I added a call of misc-perftools before
the call to the telegraf role.  I left the original (after the
call to telegraf) because calling "all" hosts will not trigger
the conditional I wrote for the file transfer.

I plan to either a) figure out a conditional that will trigger
even when passed all so I can move the call for misc-perftools
for "all" hosts into the call I put before the telegraf role call.
Or b) keep them separate but exclude compute_hosts from the later
misc-perftools call to minimize the redundancy.